--- com.unity.formats.alembic/Runtime/Scripts/Importer/AlembicCurveCollection.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 4bf0e05209b444d6ac443d7b5c8b97a2
-timeCreated: 1592848190
