+++ com.unity.formats.alembic/Runtime/Scripts/Importer/AlembicCurves.cs
@@ -1,70 +1,17 @@
+using System.Collections.Generic;
+using UnityEngine;
 using UnityEngine.Formats.Alembic.Sdk;
 
 namespace UnityEngine.Formats.Alembic.Importer
 {
-    internal class AlembicCurves : AlembicElement
+    public class AlembicCurves : MonoBehaviour
     {
-        // members
-        aiCurves m_abcSchema;
-        PinnedList<aiCurvesData> m_abcData = new PinnedList<aiCurvesData>(1);
-        aiCurvesSummary m_summary;
-        aiCurvesSampleSummary m_sampleSummary;
+        PinnedList<Vector3> m_positions = new PinnedList<Vector3>();
+        internal PinnedList<Vector3> positionsList { get { return m_positions; } }
+        public List<Vector3> Positions => positionsList.List;
 
-        internal override aiSchema abcSchema { get { return m_abcSchema; } }
-        public override bool visibility
-        {
-            get { return m_abcData[0].visibility; }
-        }
-
-        internal override void AbcSetup(aiObject abcObj, aiSchema abcSchema)
-        {
-            base.AbcSetup(abcObj, abcSchema);
-            m_abcSchema = (aiCurves)abcSchema;
-            m_abcSchema.GetSummary(ref m_summary);
-        }
-
-        public override void AbcPrepareSample()
-        {
-            base.AbcPrepareSample();
-        }
-
-        public override void AbcSyncDataBegin()
-        {
-            if (!m_abcSchema.schema.isDataUpdated)
-                return;
-
-            var sample = m_abcSchema.sample;
-            sample.GetSummary(ref m_sampleSummary);
-
-            // get points cloud component
-            var cloud = abcTreeNode.gameObject.GetComponent<AlembicCurveCollection>();
-            if (cloud == null)
-            {
-                cloud = abcTreeNode.gameObject.AddComponent<AlembicCurveCollection>();
-                //  abcTreeNode.gameObject.AddComponent<AlembicPointsRenderer>(); // Need rendering
-            }
-            var data = default(aiCurvesData);
-
-            cloud.positionsList.ResizeDiscard(m_sampleSummary.positionCount);
-            cloud.numVerticesList.ResizeDiscard(m_sampleSummary.numVerticesCount);
-            data.positions = cloud.positionsList;
-            data.numVertices = cloud.numVerticesList;
-
-            m_abcData[0] = data;
-
-            // kick async copy
-            sample.FillData(m_abcData);
-        }
-
-        public override void AbcSyncDataEnd()
-        {
-            if (!m_abcSchema.schema.isDataUpdated)
-                return;
-
-            var data = m_abcData[0];
-
-            if (abcTreeNode.stream.streamDescriptor.Settings.ImportVisibility)
-                abcTreeNode.gameObject.SetActive(data.visibility);
-        }
+        PinnedList<int> m_positionsOffsetBuffer = new PinnedList<int>();
+        internal PinnedList<int> positionOffsetBuffer { get { return m_positionsOffsetBuffer; } }
+        public List<int> PositionsOffsetBuffer => positionOffsetBuffer.List;
     }
 }
