+++ com.unity.formats.alembic/Runtime/Scripts/Importer/AlembicCurvesElement.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0b7bc2f879b54e36b7355ae5af9beaec
+guid: ce1df47d488c4484a8738ec94f9ae6ec
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
